Changes to bond creation and insert/pack
Durring bond creation, a normalized skin can be set that takes into
account the radii of the two connecting spheres. This allows a singular
bond_type to be used when bonding different sized spheres.
The insert/pack command now accepts the flag bound_insert that forces
the starting sphere of a template to a specific location.