Working code to dynamic insertion. Testing Needed
Code that dynamically inserts fibers into the simulation domain is working.
Light testing has been done against LIGGGHTS' example chute wear. More
testing is needed to make sure accurate bonds are being created and that
the bonds behave in a similar way as if you used the fix to create the bonds.