Fixed issue with highly overlapped bonded spheres
Fixed an issue with highly overlapped bond spheres. Additional code has
beend added to maintain frequency of a bonded fiber with differing
number of overlapping spheres. Previously the fiber would vibrate at a
lower frequency when adding additional spheres to the fiber to make it
"smoother". Additional code has been added to check if the normal force
should be applied with these highly overlapped spheres in a bonded
fiber. Have updated the tests to reflect this change. The biggest change
is in the value of the bond Young's modulus. If you used a value of 1.0
GPa, you should now use a value of 0.5 GPa. This also gets the value of
the bond Young's modulus closer to the contact Young's modulus (at least
with wheat straw).